Adição da página de 401, alteração de frontend e inserção de regras de negócio

+++ Controllers/API/InstrutorController.cs
@@ -4,6 +4,8 @@ using Microsoft.AspNetCore.Mvc;
 using Microsoft.EntityFrameworkCore;
 using Workshop.DB;
 using Workshop.Models;
+using System.Text.RegularExpressions;
+
 
 // For more information on enabling Web API for empty projects, visit https://go.microsoft.com/fwlink/?LinkID=397860
 
@@ -67,13 +69,30 @@ namespace Workshop.Controllers.API
         [HttpPost]
         public ActionResult Post([FromBody] Instrutor Instrutor)
         {
+            var user = HttpContext.User;
+            if (!user.HasClaim(c => c.Type == "Perfil" && c.Value == Enums.Perfil.Administrador.GetDescription()))
+                return new ObjectResult("Usuário sem permissão para cadastrar instrutores.")
+                {
+                    StatusCode = StatusCodes.Status403Forbidden
+                };
+
+
             Instrutor? InstrutorEncontrado = _context.Instrutor.Find(Instrutor.FormatCpf(Instrutor.Cpf));
             if (InstrutorEncontrado != null)
                 return BadRequest($"Um instrutor com este CPF já foi cadastrado.");
 
-            InstrutorEncontrado = _context.Instrutor.Find(Instrutor.Login);
-            if (InstrutorEncontrado != null)
+            string loginNormalizado = Instrutor.Login.Trim().ToUpper();
+
+            bool loginCadastrado = _context.Instrutor
+                .Any(i => i.Login.Trim().ToUpper() == loginNormalizado);
+            if (loginCadastrado)
                 return BadRequest($"Um instrutor com este login já foi cadastrado.");
+
+            if(Regex.Replace(Instrutor.Cpf, @"[^\d]", "").Length != 11)
+                return BadRequest($"CPF inválido. Este campo deve ter 11 caracteres.");
+
+            if (Regex.Replace(Instrutor.Telefone, @"[^\d]", "").Length != 10 && Regex.Replace(Instrutor.Telefone, @"[^\d]", "").Length != 11)
+                return BadRequest($"Telefone inválido. Este campo deve ter 10 ou 11 caracteres.");
 
             Instrutor.Senha = HashPassword(Instrutor.Senha);
             _context.Instrutor.Add(Instrutor);
@@ -84,13 +103,28 @@ namespace Workshop.Controllers.API
         [HttpPut("{cpf}")]
         public ActionResult Put(string cpf, [FromBody] Instrutor Instrutor)
         {
+            var user = HttpContext.User;
+            if (!user.HasClaim(c => c.Type == "Perfil" && c.Value == Enums.Perfil.Administrador.GetDescription()))
+                return new ObjectResult("Usuário sem permissão para editar instrutores.")
+                {
+                    StatusCode = StatusCodes.Status403Forbidden
+                };
 
-            Instrutor? InstrutorEncontrado = _context.Instrutor.Find(Instrutor.FormatCpf(cpf));
+            string cpfFormatado = Instrutor.FormatCpf(cpf);
+            Instrutor? InstrutorEncontrado = _context.Instrutor.Find(cpfFormatado);
             if (InstrutorEncontrado == null)
-                return NotFound();
+                return BadRequest($"Não foi encontrado instrutor com este CPF.");
+
+            string loginNormalizado = Instrutor.Login.Trim().ToUpper();
+
+            bool loginCadastrado = _context.Instrutor
+                .Any(i => i.Login.Trim().ToUpper() == loginNormalizado);
+            bool alteracaoLogin = !Instrutor.Login.Trim().Equals(InstrutorEncontrado.Login.Trim(), StringComparison.OrdinalIgnoreCase);
+            if (alteracaoLogin && loginCadastrado)
+                return BadRequest($"Um instrutor com este login já foi cadastrado.");
 
             if (Instrutor.Cpf == null)
-                Instrutor.Cpf = InstrutorEncontrado.Cpf;
+                Instrutor.Cpf = cpfFormatado;
 
             Instrutor.Senha = HashPassword(Instrutor.Senha);
             _context.Entry(InstrutorEncontrado).State = EntityState.Detached;
@@ -104,11 +138,20 @@ namespace Workshop.Controllers.API
         [HttpDelete("{cpf}")]
         public ActionResult Delete(string cpf)
         {
+            var user = HttpContext.User;
+            if (!user.HasClaim(c => c.Type == "Perfil" && c.Value == Enums.Perfil.Administrador.GetDescription()))
+                return new ObjectResult("Usuário sem permissão para deletar instrutores.")
+                {
+                    StatusCode = StatusCodes.Status403Forbidden
+                };
 
             Instrutor? Instrutor = _context.Instrutor.Find(Instrutor.FormatCpf(cpf));
             if (Instrutor == null)
                 return NotFound();
 
+            Models.Workshop? workshop = _context.Workshop.Include(w => w.Instrutor).FirstOrDefault(w => w.Instrutor.Cpf == Instrutor.FormatCpf(cpf));
+            if (workshop != null)
+                return BadRequest($"Não é possível deletar instrutores com workshops associados.");
 
             _context.Instrutor.Remove(Instrutor);
             _context.SaveChanges();
